part-number-extractor/requirements.txt
# Core ML Packages
torch>=2.0.0
transformers>=4.30.0
pandas>=1.5.0
numpy>=1.24.0
scikit-learn>=1.3.0
seqeval>=1.2.2
pyyaml>=6.0
tqdm>=4.65.0

# Jupyter & Visualization
jupyter>=1.0.0
matplotlib>=3.7.0
seaborn>=0.12.0

# File Handling
openpyxl>=3.1.0

# Interactive CLI Tools
colorama>=0.4.6
inquirer>=3.1.3
rich>=13.5.0
click>=8.1.0

# Optional: FastAPI (for local API server)
fastapi>=0.100.0
uvicorn>=0.23.0
python-multipart>=0.0.6

# Development & Testing
pytest>=7.4.0
black>=23.7.0
flake8>=6.1.0
